Fix error handling
In python, `sys.exit` raises a SystemExit, rather than terminating the
process, as I had thought it would. Since we invoke this from within the
context of a thread, the thread is just quietly swallowing up the
exception without passing up to the main thread.

We instead now SIGTERM the application, and print out exceptions to the
terminal, in case sentry cannot be reached.

Additionally, I made the mistake of checking to see whether or not
Sentry was configured by checking `self.sentry`, rather than using
`hasattr`. This has been fixed.